combining some tests
